Move the output registry out of WaylandState
Two fields and the three methods that maintain them: the id allocator, the
lookup back to a live `wl_output`, and the rebuild of the reactive list,
plus `OutputHandler`.

Small, but it is the piece that carries a rule worth stating in one place —
an `OutputId` is never reused, so a monitor that is unplugged and plugged
back in is a new output rather than the old one returning. A surface pinned
to the old id stays unplaced instead of silently jumping.

wayland.rs 1498 -> 1407 lines, 21 -> 20 fields.
